Updated collect employees and created new cap function
Updated collect Employees function to sort the entries by last name.
Used the sort method and compared each with greater than or equal.  This effectively means same letter or not.
Created new function to capitalize first letter of a string.
Used toUpperCase with string manipulation to only grab the first index.  Slice to keep the rest and put together.