Re-checked AC - Data validation required for just the salary.  Added.
While re-checking the AC I saw that data validation is required for just the salary (currency).  The logic dictates that it should check if its a number and additionally default to zero if no entry or a non-number is entered.  Added this logic in and tested to ensure its working.  Also added the average calculation function to include formatting for currancy.